refactor(ui): flatten `workflow-templates/` folder
- same as previous commits
- couple redundant folders here with exactly 1 file in them
