+++ ui/src/app/workflow-templates/workflow-template-container.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import {Route, RouteComponentProps, Switch} from 'react-router';
-import {WorkflowTemplateDetails} from './workflow-template-details/workflow-template-details';
-import {WorkflowTemplateList} from './workflow-template-list/workflow-template-list';
+
+import {WorkflowTemplateDetails} from './workflow-template-details';
+import {WorkflowTemplateList} from './workflow-template-list';
 
 export const WorkflowTemplateContainer = (props: RouteComponentProps<any>) => (
     <Switch>
