+++ ui/src/app/workflow-templates/workflow-template-filters.scss
