+++ ui/src/app/workflow-templates/workflow-template-link.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
-import {uiUrl} from '../../shared/base';
-import {LinkButton} from '../../shared/components/link-button';
+
+import {uiUrl} from '../shared/base';
+import {LinkButton} from '../shared/components/link-button';
 
 export const WorkflowTemplateLink = (props: {namespace: string; name: string}) => (
     <LinkButton to={uiUrl('workflow-templates/' + props.namespace + '/' + props.name)}>
