+++ ui/src/app/workflow-templates/workflow-template-list.scss
