+++ ui/src/app/workflow-templates/workflow-template-list.tsx
@@ -3,27 +3,27 @@ import * as React from 'react';
 import {useContext, useEffect, useState} from 'react';
 import {Link, RouteComponentProps} from 'react-router-dom';
 
-import {WorkflowTemplate} from '../../../../models';
-import {ANNOTATION_DESCRIPTION, ANNOTATION_TITLE} from '../../../shared/annotations';
-import {uiUrl} from '../../../shared/base';
-import {ErrorNotice} from '../../../shared/components/error-notice';
-import {ExampleManifests} from '../../../shared/components/example-manifests';
-import {InfoIcon} from '../../../shared/components/fa-icons';
-import {Loading} from '../../../shared/components/loading';
-import {PaginationPanel} from '../../../shared/components/pagination-panel';
-import {Timestamp} from '../../../shared/components/timestamp';
-import {useCollectEvent} from '../../../shared/components/use-collect-event';
-import {ZeroState} from '../../../shared/components/zero-state';
-import {Context} from '../../../shared/context';
-import {Footnote} from '../../../shared/footnote';
-import {historyUrl} from '../../../shared/history';
-import {Pagination, parseLimit} from '../../../shared/pagination';
-import {ScopedLocalStorage} from '../../../shared/scoped-local-storage';
-import {services} from '../../../shared/services';
-import {useQueryParams} from '../../../shared/use-query-params';
-import {Utils} from '../../../shared/utils';
-import {WorkflowTemplateCreator} from '../workflow-template-creator';
-import {WorkflowTemplateFilters} from '../workflow-template-filters/workflow-template-filters';
+import {WorkflowTemplate} from '../../models';
+import {ANNOTATION_DESCRIPTION, ANNOTATION_TITLE} from '../shared/annotations';
+import {uiUrl} from '../shared/base';
+import {ErrorNotice} from '../shared/components/error-notice';
+import {ExampleManifests} from '../shared/components/example-manifests';
+import {InfoIcon} from '../shared/components/fa-icons';
+import {Loading} from '../shared/components/loading';
+import {PaginationPanel} from '../shared/components/pagination-panel';
+import {Timestamp} from '../shared/components/timestamp';
+import {useCollectEvent} from '../shared/components/use-collect-event';
+import {ZeroState} from '../shared/components/zero-state';
+import {Context} from '../shared/context';
+import {Footnote} from '../shared/footnote';
+import {historyUrl} from '../shared/history';
+import {Pagination, parseLimit} from '../shared/pagination';
+import {ScopedLocalStorage} from '../shared/scoped-local-storage';
+import {services} from '../shared/services';
+import {useQueryParams} from '../shared/use-query-params';
+import {Utils} from '../shared/utils';
+import {WorkflowTemplateCreator} from './workflow-template-creator';
+import {WorkflowTemplateFilters} from './workflow-template-filters';
 
 import './workflow-template-list.scss';
